v2.2.36.17 update : half -> float

+++ PowerLit/Skybox-Procedural FarClouds.shader
@@ -340,7 +340,7 @@ SubShader {
         #if SKYBOX_SUNDISK == SKYBOX_SUNDISK_HQ
             OUT.vertex          = -v.vertex;
         #elif SKYBOX_SUNDISK == SKYBOX_SUNDISK_SIMPLE
-            OUT.rayDir          = half3(-eyeRay);
+            OUT.rayDir          = float3(-eyeRay);
         #else
             OUT.skyGroundFactor = -eyeRay.y / SKY_GROUND_THRESHOLD;
         #endif
@@ -391,15 +391,16 @@ SubShader {
         }
 
         // Calculates the sun shape
-        half calcSunAttenuation(half3 lightPos, half3 ray)
+        half calcSunAttenuation(float3 lightPos, float3 ray)
         {
         #if SKYBOX_SUNDISK == SKYBOX_SUNDISK_SIMPLE
-            half3 delta = lightPos - ray;
-            half dist = length(delta);
-            half spot = 1.0 - smoothstep(0.0, _SunSize, dist);
+            float3 delta = lightPos - ray;
+            float dist = length(delta);
+            float spot = 1.0 - smoothstep(0.0, _SunSize, dist);
             return spot * spot;
         #else // SKYBOX_SUNDISK_HQ
             half focusedEyeCos = pow(saturate(dot(lightPos, ray)), _SunSizeConvergence);
+            
             return getMiePhase(-focusedEyeCos, focusedEyeCos * focusedEyeCos);
         #endif
         }
@@ -426,15 +427,14 @@ SubShader {
             // if y >= 0 and < 1 [eyeRay.y <= 0 and > -SKY_GROUND_THRESHOLD] - horizon
             // if y < 0 [eyeRay.y > 0] - sky
             #if SKYBOX_SUNDISK == SKYBOX_SUNDISK_HQ
-                half3 ray = normalize(mul((float3x3)unity_ObjectToWorld, IN.vertex));
-                half y = ray.y / SKY_GROUND_THRESHOLD;
+                float3 ray = normalize(mul((float3x3)unity_ObjectToWorld, IN.vertex));
+                float y = ray.y / SKY_GROUND_THRESHOLD;
             #elif SKYBOX_SUNDISK == SKYBOX_SUNDISK_SIMPLE
-                half3 ray = IN.rayDir.xyz;
-                half y = ray.y / SKY_GROUND_THRESHOLD;
+                float3 ray = IN.rayDir.xyz;
+                float y = ray.y / SKY_GROUND_THRESHOLD;
             #else
-                half y = IN.skyGroundFactor;
+                float y = IN.skyGroundFactor;
             #endif
-
                 // if we did precalculate color in vprog: just do lerp between them
                 col = lerp(IN.skyColor, IN.groundColor, saturate(y));
